testsolution to show signalr singelton bug
We encountered a problem with SignalR if a hub is registered as a singleton. It works fine for a small number of requests, but as soon as we start 2 Clients which send lots of requests at once, one client successfully completes all its requests and the second client halts and waits forever for its remaining requests.

It seems like the issue was introduces with net7. If the targetframewort for the server application is changed to net6 both clients complete all their requests and everything works as expected.

I've created a small testsolution to repoduce the issue (see XXXX). This solution contians a simple server and client application. To reproduce the bug start the server and run two or mor instances of the client application at once.